intent-classifier/requirements.txt
fire
tqdm
openpyxl
pandas
scikit-learn
matplotlib
seaborn
numpy==2.1.3
tensorflow==2.19.1
tensorflow-text==2.19.0
tf_keras==2.19.0
tensorflow-hub==0.16.1
wandb==0.21.0
protobuf==5.29.5
pydantic==2.11.7
uvicorn==0.38.0
python-dotenv==1.1.1
fastapi